chore(server-side-test): modify server-side test
- modify test to check for presence of specific property value instead of just the key

[Delivers #153583784]
