+++ server/config/config.json
@@ -7,11 +7,10 @@
     "dialect": "postgres"
   },
   "test": {
-    "username": "postgres",
-    "password": "",
+    "username": "Enodi",
+    "password": null,
     "database": "postit_test",
-    "host": "127.0.0.1",
-    "port": 5432,
+    "host": "localhost",
     "dialect": "postgres",
     "logging": false
   },
