+++ server/routes/users.js
@@ -1,7 +1,6 @@
 import express from 'express';
 
 import UserController from '../controllers/UserController';
-import UserGroupController from '../controllers/UserGroupController';
 import ResetPasswordController from '../controllers/ResetPasswordController';
 import Authenticate from '../middleware/Authenticate';
 import Validations from '../middleware/Validation';
@@ -19,11 +18,11 @@ app.route('/forgotPassword')
 
 app.route('/resetPassword')
   .put(Validations.validateResetPassword,
-    ResetPasswordController.resetPassword);
+  ResetPasswordController.resetPassword);
 
 app.use(Authenticate.isLoggedIn);
 app.route('/:userId/group')
-  .get(UserGroupController.retrieveGroups);
+  .get(UserController.retrieveGroups);
 
 app.route('/search')
   .get(UserController.searchUsers);
