+++ server/test/controllers/1user.spec.js
@@ -1,8 +1,8 @@
 import supertest from 'supertest';
 import { expect } from 'chai';
 
-import app from '../../../index';
 import db from '../../models';
+import app from '../../../index';
 
 const request = supertest(app);
 let token;
@@ -31,9 +31,10 @@ describe('POST /api/v1/user/signup', () => {
           expect(response.status).to.equal(201);
           expect(response.body.message).to.equal('Signup successful');
           expect(response.body.username).to.equal('enodi');
+          expect(response.body.email).to.equal('enodiaudu5@gmail.com');
+          expect(response.body.fullname).to.equal('Enodi Audu');
           expect(response.body).to.be.an('object');
-          expect(response.body).to.have.property('message')
-            .but.not.property('password');
+          expect(response.body).to.have.property('message');
           done();
         });
     });
@@ -51,6 +52,8 @@ describe('POST /api/v1/user/signup', () => {
           expect(response.status).to.equal(201);
           expect(response.body.message).to.equal('Signup successful');
           expect(response.body.username).to.equal('agnes');
+          expect(response.body.email).to.equal('agnes@gmail.com');
+          expect(response.body.fullname).to.equal('Agnes Audu');
           expect(response.body).to.be.an('object');
           expect(response.body).to.have.property('username')
             .but.not.property('password');
@@ -127,7 +130,7 @@ describe('POST /api/v1/user/signup', () => {
         })
         .end((err, response) => {
           expect(response.status).to.equal(400);
-          expect(typeof response.body.message).to.equal('object');
+          expect(response.body.message).to.be.an('object');
           expect(response.body.message).to.have.property('email');
           expect(response.body.message.email).to.be.an('array');
           expect(response.body.message.email[0])
@@ -146,7 +149,7 @@ describe('POST /api/v1/user/signup', () => {
         })
         .end((err, response) => {
           expect(response.status).to.equal(400);
-          expect(typeof response.body.message).to.equal('object');
+          expect(response.body.message).to.be.an('object');
           expect(response.body.message).to.have.property('fullname');
           expect(response.body.message.fullname).to.be.an('array');
           expect(response.body.message.fullname[0])
@@ -165,7 +168,7 @@ describe('POST /api/v1/user/signup', () => {
         })
         .end((err, response) => {
           expect(response.status).to.equal(400);
-          expect(typeof response.body.message).to.equal('object');
+          expect(response.body.message).to.be.an('object');
           expect(response.body.message).to.have.property('password');
           expect(response.body.message.password).to.be.an('array');
           expect(response.body.message.password[0])
@@ -182,6 +185,14 @@ describe('POST /api/v1/user/signup', () => {
           .end((err, response) => {
             expect(response.status).to.equal(400);
             expect(response.body.message).to.be.an('object');
+            expect(response.body.message.username[0])
+              .to.equal('The username field is required.');
+            expect(response.body.message.email[0])
+              .to.equal('The email field is required.');
+            expect(response.body.message.password[0])
+              .to.equal('The password field is required.');
+            expect(response.body.message.fullname[0])
+              .to.equal('The fullname field is required.');
             expect(response.body.message)
               .to.have.all.keys('username', 'email', 'password', 'fullname');
             expect(response.body.message.password).to.be.an('array');
@@ -208,7 +219,7 @@ describe('POST /api/v1/user/signup', () => {
           done();
         });
     });
-    
+
     it('should return 400 when user provides username with less than 4 characters',
       (done) => {
         request
@@ -293,8 +304,8 @@ describe('POST /api/v1/user/signup', () => {
           });
       });
 
-    it(`should return 400 when user passes whitespace
-    into password field`, (done) => {
+    it('should return 400 when user passes whitespace into password field',
+      (done) => {
         request
           .post('/api/v1/user/signup')
           .send({
@@ -330,6 +341,7 @@ describe('POST /api/v1/user/signin', () => {
           expect(response.body.success).to.equal(true);
           expect(response.body).to.be.an('object');
           expect(response.body).to.have.all.keys('success', 'token');
+          token = response.body.token;
           done();
         });
     });
@@ -398,21 +410,76 @@ describe('POST /api/v1/user/signin', () => {
   });
 });
 
+describe('GET /api/v1/user/:userId/group', () => {
+  describe('handles retrieving groups a user belongs to', () => {
+    it('should return 201 when a group is created successfully', (done) => {
+      request
+        .post('/api/v1/group')
+        .set('x-access-token', token)
+        .send({
+          name: 'general',
+          description: 'general channel'
+        })
+        .end((err, response) => {
+          expect(response.status).to.equal(201);
+          expect(response.body).to.be.an('object');
+          expect(response.body.message).to.equal('Group created successfully');
+          expect(response.body.id).to.equal(1);
+          expect(response.body.name).to.equal('general');
+          expect(response.body.description).to.equal('general channel');
+          expect(response.body).to.have.all
+            .keys('id', 'message', 'name', 'description', 'createdAt');
+          done();
+        });
+    });
 
-describe('GET /api/v1/user/search', () => {
-  before((done) => {
-    request
-      .post('/api/v1/user/signin')
-      .send({
-        username: 'enodi',
-        password: 'password',
-      })
-      .end((err, response) => {
-        token = response.body.token;
-        done();
+    it('should return 200 when group is retrieved successfully', (done) => {
+      request
+        .get(`/api/v1/user/${1}/group`)
+        .set('x-access-token', token)
+        .end((err, response) => {
+          expect(response.status).to.equal(200);
+          expect(response.body).to.be.an('object');
+          expect(response.body).to.have.property('groups');
+          expect(response.body.groups.id).to.equal(1);
+          expect(response.body.groups).to.be.an('object');
+          expect(response.body.groups)
+            .to.have.all.keys('id', 'email', 'username', 'fullname', 'Groups');
+          done();
+        });
+    });
+
+    it('should return 400 when a wrong value is passed', (done) => {
+      request
+        .get(`/api/v1/user/${' '}/group`)
+        .set('x-access-token', token)
+        .end((err, response) => {
+          expect(response.status).to.equal(400);
+          expect(response.body).to.be.an('object');
+          expect(response.body).to.have.property('error');
+          expect(response.body.error).to.equal('Invalid User Id');
+          done();
+        });
+    });
+
+    it('should return 401 when an unauthenticated user tries to access this route',
+      (done) => {
+        request
+          .get('/api/v1/user/:userId/group')
+          .end((err, response) => {
+            expect(response.status).to.equal(401);
+            expect(response.body).to.be.an('object');
+            expect(response.body.message)
+              .to.equal('Access denied, Authentication token does not exist');
+            expect(response.body.success).to.equal(false);
+            expect(response.body).to.have.all.keys('success', 'message');
+            done();
+          });
       });
   });
+});
 
+describe('GET /api/v1/user/search', () => {
   describe('handles retrieving users', () => {
     it('should return 200 when user is retrieved successfully', (done) => {
       request
@@ -422,6 +489,10 @@ describe('GET /api/v1/user/search', () => {
         .end((err, response) => {
           expect(response.status).to.equal(200);
           expect(response.body).to.be.an('object');
+          expect(response.body.rows[0].id).to.equal(1);
+          expect(response.body.rows[0].username).to.equal('enodi');
+          expect(response.body.rows[0].fullname).to.equal('Enodi Audu');
+          expect(response.body.rows[0].email).to.equal('enodiaudu5@gmail.com');
           expect(response.body).to.have.all.keys('count', 'rows');
           expect(response.body.rows).to.be.an('array');
           expect(response.body.rows[0])
@@ -464,193 +535,14 @@ describe('GET /api/v1/user/search', () => {
         .end((err, response) => {
           expect(response.status).to.equal(401);
           expect(response.body).to.be.an('object');
-          expect(response.body.message).to.equal('Failed to Authenticate Token');
+          expect(response.body.message).to
+            .equal('Failed to Authenticate Token');
           expect(response.body.success).to.equal(false);
           expect(response.body).to.have.all.keys('success', 'message', 'error');
           expect(response.body.error).to.be.an('object');
           expect(response.body.error).to.have.all.keys('name', 'message');
-          done();
-        });
-    });
-  });
-});
-
-describe('POST /api/v1/group/:groupId/user', () => {
-  before((done) => {
-    request
-      .post('/api/v1/user/signin')
-      .send({
-        username: 'enodi',
-        password: 'password',
-      })
-      .end((err, response) => {
-        token = response.body.token;
-        done();
-      });
-  });
-  describe('handles adding users to group', () => {
-    it('should return 401 when an unauthenticated user tries to access this route',
-      (done) => {
-        request
-          .post('/api/v1/group/:groupId/user')
-          .end((err, response) => {
-            expect(response.status).to.equal(401);
-            expect(response.body).to.be.an('object');
-            expect(response.body.message)
-              .to.equal('Access denied, Authentication token does not exist');
-            expect(response.body.success).to.equal(false);
-            expect(response.body).to.have.all.keys('success', 'message');
-            done();
-          });
-      });
-    it('should return 201 when a group is created successfully', (done) => {
-      request
-        .post('/api/v1/group')
-        .set('x-access-token', token)
-        .send({
-          name: 'andela',
-          description: 'andela fellows'
-        })
-        .end((err, response) => {
-          expect(response.status).to.equal(201);
-          expect(response.body).to.be.an('object');
-          expect(response.body)
-            .to.have.all.keys('id', 'name', 'description', 'createdAt', 'message');
-          expect(response.body.message).to.equal('Group created successfully');
-          done();
-        });
-    });
-    it('should return 404 when user tries to add a user that doesn\'t exist',
-      (done) => {
-        request
-          .post(`/api/v1/group/${1}/user`)
-          .set('x-access-token', token)
-          .send({
-            userId: 10
-          })
-          .end((err, response) => {
-            expect(response.status).to.equal(404);
-            expect(response.body).to.be.an('object');
-            expect(response.body).to.have.property('message');
-            expect(response.body.message).to.equal('User not found');
-            done();
-          });
-      });
-    it('should return 409 when user tries to add an existing user', (done) => {
-      request
-        .post(`/api/v1/group/${1}/user`)
-        .set('x-access-token', token)
-        .send({
-          userId: 1
-        })
-        .end((err, response) => {
-          expect(response.status).to.equal(409);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.property('message');
-          expect(response.body.message)
-            .to.equal('User is already a member of the group');
-          done();
-        });
-    });
-    
-    it('should return 404 when group doesn\'t exist', (done) => {
-      request
-        .post(`/api/v1/group/${10}/user`)
-        .set('x-access-token', token)
-        .send({
-          userId: 1
-        })
-        .end((err, response) => {
-          expect(response.status).to.equal(404);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.property('message');
-          expect(response.body.message).to.equal('Group doesn\'t exist');
-          done();
-        });
-    });
-    it('should return 200 when user is added to group', (done) => {
-      request
-        .post(`/api/v1/group/${1}/user`)
-        .set('x-access-token', token)
-        .send({
-          userId: 2
-        })
-        .end((err, response) => {
-          expect(response.status).to.equal(200);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.property('message');
-          expect(response.body.message).to.equal('User added successfully');
-          done();
-        });
-    });
-  });
-});
-
-describe('GET /api/v1/group/:groupId/users', () => {
-  before((done) => {
-    request
-      .post('/api/v1/user/signin')
-      .send({
-        username: 'enodi',
-        password: 'password',
-      })
-      .end((err, response) => {
-        token = response.body.token;
-        done();
-      });
-  });
-  describe('handles retrieving users from group', () => {
-    it('should return 401 when an unauthenticated user tries to access this route',
-      (done) => {
-        request
-          .get(`/api/v1/group/${1}/users`)
-          .end((err, response) => {
-            expect(response.status).to.equal(401);
-            expect(response.body).to.be.an('object');
-            expect(response.body.message)
-              .to.equal('Access denied, Authentication token does not exist');
-            expect(response.body.success).to.equal(false);
-            expect(response.body).to.have.all.keys('success', 'message');
-            done();
-          });
-      });
-    it('should return 400 when no groupId is passed', (done) => {
-      request
-        .get('/api/v1/group/:groupId/users')
-        .set('x-access-token', token)
-        .end((err, response) => {
-          expect(response.status).to.equal(400);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.property('message');
-          expect(response.body.message).to.equal('Please specify a groupId');
-          done();
-        });
-    });
-    it('should return 200 when users are retrieved successfully', (done) => {
-      request
-        .get(`/api/v1/group/${1}/users`)
-        .set('x-access-token', token)
-        .end((err, response) => {
-          expect(response.status).to.equal(200);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.all.keys('groupUsers', 'message');
-          expect(response.body.groupUsers).to.be.an('array');
-          expect(response.body.groupUsers[0])
-            .to.have.all.keys('id', 'email', 'password',
-            'username', 'fullname', 'createdAt', 'updatedAt', 'UserGroup');
-          expect(response.body.message).to.equal('Users retrieved successfully');
-          done();
-        });
-    });
-    it('should return 404 when group doesn\'t exist', (done) => {
-      request
-        .get(`/api/v1/group/${10}/users`)
-        .set('x-access-token', token)
-        .end((err, response) => {
-          expect(response.status).to.equal(404);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.property('message');
-          expect(response.body.message).to.equal('Group doesn\'t exist');
+          expect(response.body.error.name).to.equal('JsonWebTokenError');
+          expect(response.body.error.message).to.equal('jwt malformed');
           done();
         });
     });
