+++ server/test/controllers/3message.spec.js
@@ -9,39 +9,47 @@ let token;
 describe('POST /api/v1/group/:groupId/message', () => {
   before((done) => {
     request
-      .post('/api/v1/user/signup')
+      .post('/api/v1/user/signin')
       .send({
-        username: 'enodi000',
-        email: 'enodiaudu5@gmail.co.uk',
-        fullname: 'Enodi Audu',
+        username: 'enodi',
         password: 'password',
       })
       .end((err, response) => {
+        expect(response.status).to.equal(200);
+        expect(response.body.success).to.equal(true);
+        expect(response.body).to.be.an('object');
+        expect(response.body).to.have.all.keys('success', 'token');
         token = response.body.token;
         done();
       });
   });
   describe('handles posting messages', () => {
-    it('should return 201 when a message is posted with priority critical', (done) => {
-      request
-        .post(`/api/v1/group/${1}/message`)
-        .set('x-access-token', token)
-        .send({
-          message: 'hello world',
-          priority: 'critical'
-        })
-        .end((err, response) => {
-          expect(response.status).to.equal(201);
-          expect(response.body).to.be.an('object');
-          expect(response.body).to.have.all.keys('messageCreated', 'message');
-          expect(response.body.messageCreated).to.be.an('object');
-          expect(response.body.messageCreated)
-            .to.have.all.keys('id', 'UserId', 'GroupId',
-            'message', 'priority', 'updatedAt', 'createdAt');
-          expect(response.body.message).to.equal('message posted successfully');
-          done();
-        });
-    });
+    it('should return 201 when a message is posted with priority critical',
+      (done) => {
+        request
+          .post(`/api/v1/group/${1}/message`)
+          .set('x-access-token', token)
+          .send({
+            message: 'hello world',
+            priority: 'critical'
+          })
+          .end((err, response) => {
+            expect(response.status).to.equal(201);
+            expect(response.body).to.be.an('object');
+            expect(response.body.message).to
+              .equal('message posted successfully');
+            expect(response.body.messageCreated.id).to.equal(1);
+            expect(response.body.messageCreated.message).to
+              .equal('hello world');
+            expect(response.body.messageCreated.priority).to.equal('critical');
+            expect(response.body).to.have.all.keys('messageCreated', 'message');
+            expect(response.body.messageCreated).to.be.an('object');
+            expect(response.body.messageCreated)
+              .to.have.all.keys('id', 'UserId', 'GroupId',
+              'message', 'priority', 'updatedAt', 'createdAt');
+            done();
+          });
+      });
 
     it('should return 201 when a message is posted with priority urgent',
       (done) => {
@@ -55,12 +63,17 @@ describe('POST /api/v1/group/:groupId/message', () => {
           .end((err, response) => {
             expect(response.status).to.equal(201);
             expect(response.body).to.be.an('object');
+            expect(response.body.message).to
+              .equal('message posted successfully');
+            expect(response.body.messageCreated.id).to.equal(2);
+            expect(response.body.messageCreated.message).to
+              .equal('hello');
+            expect(response.body.messageCreated.priority).to.equal('urgent');
             expect(response.body).to.have.all.keys('messageCreated', 'message');
             expect(response.body.messageCreated).to.be.an('object');
             expect(response.body.messageCreated)
               .to.have.all.keys('id', 'UserId', 'GroupId',
               'message', 'priority', 'updatedAt', 'createdAt');
-            expect(response.body.message).to.equal('message posted successfully');
             done();
           });
       });
@@ -124,6 +137,9 @@ describe('GET /api/v1/group/:groupId/messages', () => {
           expect(response.status).to.equal(200);
           expect(response.body).to.be.an('object');
           expect(response.body.message).to.equal('messages retrieved');
+          expect(response.body.messageRetrieved[0].id).to.equal(2);
+          expect(response.body.messageRetrieved[0].message).to.equal('hello');
+          expect(response.body.messageRetrieved[0].priority).to.equal('urgent');
           expect(response.body).to.have.all.keys('messageRetrieved', 'message');
           expect(response.body.messageRetrieved).to.be.an('array');
           expect(response.body.messageRetrieved[0])
